// frontend/src/shared/api/generated/model/roadmapSectionOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapLessonOut } from './roadmapLessonOut';
import type { RoadmapSubsectionOut } from './roadmapSubsectionOut';

export interface RoadmapSectionOut {
  id: string;
  name: string;
  subsections: RoadmapSubsectionOut[];
  lessons: RoadmapLessonOut[];
}